getNotification ById now works
